Use custom errors and improve API
Version 2.0.0 (#33)

This PR uses `thiserror` instead of `anyhow` with custom errors to improve the error handling. It also modify the API to make it a bit more usable, replacing `Result` with `Option` when it's more natural and other changes like that.

This PR also clean the code and make it a bit more readable.