Change the type of metadata to Vec<MetadataItem>
Struct `MetadataItem` is introduced mainly to support more information
than the data value (e.g,. refinements such as "file-as").

The reason to change from HashMap to Vec isn't significant.
Vec is slightly easier to construct and does reflect the original
element better, preserving the data items' order.

In this commit, metadata does not have more or less info.
The incomplete construction for EPUB2 will be completed next,
and the implementation for EPUB3 following that.